+++ src/components/MissionList/HeroList/HeroItem.module.css
