+++ src/components/MissionList/HeroList/HeroItem.tsx
