+++ src/components/MissionList/MissionCard/MissionCard.tsx
