+++ src/components/MissionList/MissionInfo/MissionInfo.tsx
