+++ src/components/MissionList/MissionItem/MissionItem.tsx
