+++ src/components/MissionList/Result/Result.module.css
