+++ src/components/MissionList/Result/Result.tsx
