+++ src/components/MissionList/Result/index.ts
