+++ src/pages/api/groups/[groupId]/missionlist/[missionId].ts
