+++ src/pages/missionlist/index.tsx
@@ -8,15 +8,14 @@ import { useRecoilValue } from 'recoil';
 import { filteredMissionListState } from '@/states';
 import { useDialog, useHeroes, useMissions } from '@/hooks';
 import { Dialog, Nav, Slide } from '@/components/common';
-import { MissionCard, MissionItem } from '@/components/Missions';
-import styles from '@styles/Missions.module.css';
+import { MissionCard, MissionItem } from '@/components/MissionList';
+import styles from '@styles/MissionList.module.css';
 
-export default function Missions() {
+export default function MissionList() {
   const router = useRouter();
   const { initHeroeList, getHero } = useHeroes();
   const { initMissionList } = useMissions();
   const filteredMissionList = useRecoilValue(filteredMissionListState);
-  const missionDialog = useDialog();
   const [selectedMission, setSelectedMission] = useState<Mission | null>(null);
 
   useEffect(() => {
@@ -24,29 +23,13 @@ export default function Missions() {
     initMissionList();
   }, []);
 
-  const openMissionInfo = (mission: Mission) => {
-    setSelectedMission(mission);
-    missionDialog.open();
-  };
-  const openMissionForm = () => {
-    missionDialog.open();
-  };
-  const closeMissionCard = () => {
-    setSelectedMission(null);
-    missionDialog.close();
-  };
-
   return (
     <>
       <Head>
         <title>임무를 수행해보세요!</title>
       </Head>
       <div className={styles.container}>
-        <Nav
-          buttonName="임무 등록"
-          onButtonClick={openMissionForm}
-          currentPage={router?.asPath}
-        />
+        <Nav buttonName="임무 등록" currentPage={router?.asPath} />
         <Slide direction="left" className={styles.missionContainer}>
           {filteredMissionList.length > 0 ? (
             <ul className={styles.missionList}>
@@ -56,7 +39,6 @@ export default function Missions() {
                     key={mission.id}
                     author={getHero(mission.authorId) as Hero}
                     mission={mission}
-                    onClick={openMissionInfo}
                   />
                 );
               })}
@@ -65,14 +47,6 @@ export default function Missions() {
             <p>임무를 등록해주세요!</p>
           )}
         </Slide>
-        {missionDialog.isOpen ? (
-          <Dialog modal onClose={closeMissionCard}>
-            <MissionCard
-              initialMission={selectedMission}
-              onClose={closeMissionCard}
-            />
-          </Dialog>
-        ) : null}
       </div>
     </>
   );
